add the first version.
this contains a basic walkthrough that should be pretty complete
regarding typescripts features.
there's a lot of room for improvement regarding the coding sections.

i chose brunch as the setup is really amazingly simple and we do not
want to be distracted here by a fancy webpack/rollup/gulp/whatnot-setup.